requirements.txt
textual==1.0.0
numpy==2.0.2
pyaudio==0.2.14
pydub==0.25.1
sounddevice==0.5.1
python-dotenv==1.0.1
uvicorn==0.34.0
openai[realtime]
fastapi==0.115.8